+++ bitbucket.php
@@ -1,27 +1,31 @@
 <?php
-
-/*{{{ v.151005.001 (0.0.2)
-
-    Routines for work with bitbucket server, repositories and projects.
-
-    Based on 'Automated git deployment' script by Jonathan Nicoal:
-    http://jonathannicol.com/blog/2013/11/19/automated-git-deployments-from-bitbucket/
-
-    See README.md and config.sample.php
-
-    ---
-    Igor Lilliputten
-    mailto: igor at lilliputten dot ru
-    http://lilliputtem.ru/
-
-    Ivan Pushkin
-    mailto: iv dot pushk at gmail dot com
-
-}}}*/
+/**
+ * @module bitbucket
+ * @version 2018.10.21, 02:50
+ *
+ * Routines for work with bitbucket server, repositories and projects.
+ *
+ * Based on 'Automated git deployment' script by Jonathan Nicoal:
+ * http://jonathannicol.com/blog/2013/11/19/automated-git-deployments-from-bitbucket/
+ *
+ * See README.md and config.sample.php
+ *
+ * ---
+ * Igor Lilliputten
+ * mailto: igor at lilliputten dot ru
+ * http://lilliputten.ru/
+ *
+ * Ivan Pushkin
+ * mailto: iv dot pushk at gmail dot com
+ */
 
 /*{{{ *** Global variables */
 
 define('DEFAULT_FOLDER_MODE', 0755);
+
+if ( !defined('NL') ) {
+    define('NL',"\n");
+}
 
 $PAYLOAD   = array ();
 $BRANCHES  = array ();
@@ -32,16 +36,31 @@ $REPO_NAME = ''; // name
 
 function initConfig ()/*{{{ Initializing repo configs */
 {
-    global $PROJECTS;
+    global $CONFIG, $PROJECTS;
 
     $tmpProjects = array();
 
     // Bitbucket uses lower case repo names!
+    $hadUppercaseKeys = false;
     foreach ( $PROJECTS as $repoName => $config ) {
         $tmpProjects[strtolower($repoName)] = $config;
+        $hadUppercaseKeys = true;
     }
 
-    $PROJECTS = $tmpProjects;
+    // Rewrite projects list if has changes
+    if ( $hadUppercaseKeys ) {
+        $PROJECTS = $tmpProjects;
+    }
+
+    // Set default folder mode if absent
+    if ( empty($CONFIG['folderMode']) ) {
+        $CONFIG['folderMode'] = DEFAULT_FOLDER_MODE;
+    }
+
+    if ( $CONFIG['verbose'] ) {
+        _LOG_VAR('CONFIG',$CONFIG);
+    }
+
 }/*}}}*/
 function initLog ()/*{{{ Initializing log variables */
 {
@@ -60,7 +79,7 @@ function initLog ()/*{{{ Initializing log variables */
 }/*}}}*/
 function initPayload ()/*{{{ Get posted data */
 {
-    global $PAYLOAD;
+    global $CONFIG, $PAYLOAD;
 
     if (isset($_SERVER['HTTP_X_EVENT_KEY'], $_SERVER['HTTP_X_HOOK_UUID'], $_SERVER['HTTP_USER_AGENT'],
         $_SERVER['REMOTE_ADDR'])) {
@@ -87,33 +106,40 @@ function initPayload ()/*{{{ Get posted data */
         exit;
     }
 
-    _LOG("Valid payload was received");
+    _LOG("Valid payload was received!");
+    if ( $CONFIG['logPayload'] ) {
+        _LOG_VAR('PAYLOAD',$PAYLOAD);
+    }
 
 }/*}}}*/
 function fetchParams ()/*{{{ Get parameters from bitbucket payload now only (REPO) */
 {
-    global $REPO, $REPO_NAME, $PAYLOAD, $PROJECTS, $BRANCHES;
+    global $CONFIG, $REPO, $REPO_NAME, $PAYLOAD, $PROJECTS, $BRANCHES;
 
     // Get repository name:
     $REPO = strtolower($PAYLOAD->repository->full_name);
+    // _DEBUG_VAR('REPO', $REPO);
     if ( empty($PROJECTS[$REPO]) ) {
         _ERROR("Not found repository config for '$REPO'!");
         exit;
     }
 
     $REPO_NAME = strtolower($PAYLOAD->repository->name);
+    _LOG_VAR('Repository name', $REPO_NAME);
 
     foreach ( $PAYLOAD->push->changes as $change ) {
-        if ( is_object($change->new) && $change->new->type == "branch" &&
-            isset($PROJECTS[$REPO][$change->new->name]) ) {
-            // Create branch name for checkout
-            array_push($BRANCHES, $change->new->name);
-            _LOG("Changes in branch '".$change->new->name."' was fetched");
+        if ( is_object($change->new) && $change->new->type == "branch" ) {
+            $branchName = $change->new->name;
+            if ( isset($PROJECTS[$REPO][$branchName]) ) {
+                // Create branch name for checkout
+                array_push($BRANCHES, $branchName);
+                _LOG("Changes in branch '".$branchName."' was fetched");
+            }
         }
     }
 
     if ( empty($BRANCHES) ) {
-        _LOG("Nothing to update");
+        _ERROR("Nothing to update (no branches found)! Please check correct branch names in your config PROJECTS list.");
     }
 
 }/*}}}*/
@@ -122,30 +148,32 @@ function checkPaths ()/*{{{ Check repository and project paths; create them if n
     global $REPO, $CONFIG, $PROJECTS, $BRANCHES;
 
     // Check for repositories folder path; create if absent
-    if ( !is_dir($CONFIG['repositoriesPath']) ) {
-        $mode = ( !empty($CONFIG['folderMode']) ) ? $CONFIG['folderMode'] : DEFAULT_FOLDER_MODE;
-
-        if ( mkdir($CONFIG['repositoriesPath'],$mode,true) ) {
-            _LOG("Creating repository folder '".$CONFIG['repositoriesPath']." (".decoct($mode).") for '$REPO'");
+    $repositoriesPath = $CONFIG['repositoriesPath'];
+    if ( !is_dir($repositoriesPath) ) {
+        $mode = $CONFIG['folderMode'];
+        if ( mkdir($repositoriesPath,$mode,true) ) {
+            chmod($repositoriesPath,$mode); // NOTE: Ensuring folder mode!
+            _LOG("Creating repository folder '".$repositoriesPath." (".decoct($mode).") for '$REPO'");
         }
         else {
-            _ERROR("Error creating repository folder '".$CONFIG['repositoriesPath']." for '$REPO'! Exiting.");
+            _ERROR("Error creating repository folder '".$repositoriesPath." for '$REPO'! Exiting.");
             exit;
         }
     }
 
     // Create folder if absent for each pushed branch
     foreach ( $BRANCHES as $branchName ) {
-        if ( !is_dir($PROJECTS[$REPO][$branchName]['deployPath']) ) {
-            $mode = ( !empty($CONFIG['folderMode']) ) ? $CONFIG['folderMode'] : DEFAULT_FOLDER_MODE;
-
-            if ( mkdir($PROJECTS[$REPO][$branchName]['deployPath'],$mode,true) ) {
-                _LOG("Creating project folder '".$PROJECTS[$REPO][$branchName]['deployPath'].
+        $deployPath = $PROJECTS[$REPO][$branchName]['deployPath'];
+        if ( !is_dir($deployPath) ) {
+            $mode = $CONFIG['folderMode'];
+            if ( mkdir($deployPath,$mode,true) ) {
+                chmod($deployPath,$mode); // NOTE: Ensuring folder mode!
+                _LOG("Creating project folder '".$deployPath.
                     " (".decoct($mode).") for '$REPO' branch '$branchName'");
             }
             else {
-                _ERROR("Error creating project folder '".$PROJECTS[$REPO][$branchName]['deployPath'].
-                    " for '$REPO' branch '$branchName'! Exiting.");
+                _ERROR("Error creating project folder '".$deployPath.
+                    "' for '$REPO' branch '$branchName'! Exiting.");
                 exit;
             }
         }
@@ -157,10 +185,9 @@ function placeVerboseInfo ()/*{{{ Place verbose log information -- if specified 
     global $REPO, $REPO_NAME, $CONFIG, $BRANCHES;
 
     if ( $CONFIG['verbose'] ) {
-        _LOG_VAR('CONFIG',$CONFIG);
-        _LOG_VAR('REPO',$REPO);
+        // _LOG_VAR('REPO',$REPO);
         _LOG_VAR('repoPath',$CONFIG['repositoriesPath'].DIRECTORY_SEPARATOR.$REPO_NAME.'.git');
-        _LOG_VAR('BRANCHES',$BRANCHES);
+        // _LOG_VAR('BRANCHES',$BRANCHES);
     }
 }/*}}}*/
 function fetchRepository ()/*{{{ Fetch or clone repository */
@@ -172,28 +199,30 @@ function fetchRepository ()/*{{{ Fetch or clone repository */
 
     // If repository or repository folder are absent then clone full repository
     if ( !is_dir($repoPath) || !is_file($repoPath.DIRECTORY_SEPARATOR.'HEAD') ) {
-        _LOG("Absent repository for '$REPO', cloning");
+        _LOG("Repository folder absent for '$REPO', cloning...");
 
         $cmd = 'cd '.$CONFIG['repositoriesPath'].' && '.$CONFIG['gitCommand'].
-            ' clone --mirror git@bitbucket.org:'.$REPO.'.git';
+            ' clone --mirror git@bitbucket.org:'.$REPO.'.git 2>&1';
         _LOG_VAR('cmd',$cmd);
-        system($cmd, $status);
+        // system($cmd, $status);
+        exec($cmd, $output, $status);
 
         if ( $status !== 0 ) {
-            _ERROR('Cannot clone repository git@bitbucket.org:'.$REPO.'.git');
+            _ERROR('Cannot clone repository git@bitbucket.org:'.$REPO.'.git: '.NL.implode(NL,$output));
             exit;
         }
     }
     // Else fetch changes
     else {
-        _LOG("Fetching repository '$REPO'");
+        _LOG("Repositury folder exists for '$REPO', fetching...");
 
-        $cmd = 'cd '.$repoPath.' && '.$CONFIG['gitCommand'].' fetch';
+        $cmd = 'cd '.$repoPath.' && '.$CONFIG['gitCommand'].' fetch 2>&1';
         _LOG_VAR('cmd',$cmd);
-        system($cmd, $status);
+        // system($cmd, $status);
+        exec($cmd, $output, $status);
 
         if ( $status !== 0 ) {
-            _ERROR("Cannot fetch repository '$REPO' in '$repoPath'!");
+            _ERROR("Cannot fetch repository '$REPO' in '$repoPath': ".NL.implode(NL,$output));
             exit;
         }
     }
@@ -211,20 +240,22 @@ function checkoutProject ()/*{{{ Checkout project into target folder */
         $cmd = 'cd '.$repoPath.' && GIT_WORK_TREE='.$PROJECTS[$REPO][$branchName]['deployPath']
             .' '.$CONFIG['gitCommand'].' checkout -f '.$branchName;
         _LOG_VAR('cmd',$cmd);
-        system($cmd, $status);
+        // system($cmd, $status);
+        exec($cmd, $output, $status);
 
         if ( $status !== 0 ) {
-            _ERROR("Cannot checkout branch '$branchName' in repo '$REPO'!");
+            _ERROR("Cannot checkout branch '$branchName' in repo '$REPO': ".NL.implode(NL,$output));
             exit;
         }
 
         if ( !empty($PROJECTS[$REPO][$branchName]['postHookCmd']) ) {
             $cmd = 'cd '.$PROJECTS[$REPO][$branchName]['deployPath'].' && '.$PROJECTS[$REPO][$branchName]['postHookCmd'];
             _LOG_VAR('cmd',$cmd);
-            system($cmd, $status);
+            // system($cmd, $status);
+            exec($cmd, $output, $status);
 
             if ( $status !== 0 ) {
-                _ERROR("Error in post hook command for branch '$branchName' in repo '$REPO'!");
+                _ERROR("Error in post hook command for branch '$branchName' in repo '$REPO': ".NL.implode(NL,$output));
                 exit;
             }
         }