+++ CHANGELOG.md
@@ -3,4 +3,7 @@
 ## v.0.1.1
 
 - Verbose exec errors (show error messages in log).
+- Ensuring folder modes after `mkdir` (calling `chmod`). In some cases mkdir makes default modes.
+- More informative log messages.
+- New config parameters: `logDebug`, `logPayload`.